refactor(core): fix layer violations and improve module abstraction
- Remove unused getConfig from ResolverHandle, eliminating kernel -> built-in-service reverse dependency
- Move createTestTargetBase from kernel to app layer where it belongs
- Move config-token integration test to app layer
- Remove unnecessary token re-exports from app/index.ts
- Refactor default-modules to loop over Module interface via bivariant widening

Verified: precommit-verify ✓ (ee2572b27bc4c8c8)
