+++ packages/core/src/app/config-token.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it } from 'vitest';
 
-import { createApp } from '../../app';
-import { inject } from '../../kernel/di/inject';
-import { Controller } from '../../modules/http/decorators/controller';
-import { Get } from '../../modules/http/decorators/http-method';
-import { Config } from './decorator';
+import { Config } from '../built-in-service/config';
+import { inject } from '../kernel/di/inject';
+import { Controller } from '../modules/http/decorators/controller';
+import { Get } from '../modules/http/decorators/http-method';
+import { createApp } from './create-app';
 
 describe('config token', () => {
   describe('single @Config', () => {
