+++ packages/core/src/app/index.ts
@@ -1,6 +1,3 @@
-export { COMMAND_OPTIONS } from '../modules/command/module';
-export { HTTP_OPTIONS } from '../modules/http/module';
-export { SCHEDULER_OPTIONS } from '../modules/scheduler/module';
 export { AppRuntime } from './app-runtime';
 export { ConfigRegistry } from './config-registry';
 export {
