+++ packages/core/src/built-in-service/logger/logger.service.test.ts
@@ -1,5 +1,5 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
-import { createTestTargetBase } from '../../kernel/di/container';
+import { createTestTargetBase } from '../../app/test-target';
 import { Config } from '../config';
 
 import { LoggerConfig } from './logger.config';
