+++ packages/core/src/internal-bridge/testing.ts
@@ -1,2 +1,2 @@
-export type { CreateTestTargetOptions, TestTargetResult } from '../kernel/di/container';
-export { createTestTargetBase } from '../kernel/di/container';
+export type { CreateTestTargetOptions, TestTargetResult } from '../app/test-target';
+export { createTestTargetBase } from '../app/test-target';
